Change Signal::operator() to no longer call on expired slots.
This was previously expected behaviour, for a Signal to call expired
slots and an exception being throw. It is not exceptional behaviour
when a Signal is called with an expired slot. Signals should be
expected to no longer call on expired slots, this is the whole point
of Slot::track.